lib-unlimited/src/test/java/fr/tduf/libunlimited/common/helper/FilesHelperTest.java
package fr.tduf.libunlimited.common.helper;

import org.junit.Before;
import org.junit.Test;

import java.io.File;
import java.io.IOException;
import java.nio.file.Files;

import static org.assertj.core.api.Assertions.assertThat;

public class FilesHelperTest {

    private String tempDirectory;

    @Before
    public void setUp() throws IOException {
        tempDirectory = Files.createTempDirectory("libUnlimited-tests").toString();
    }

    @Test
    public void createDirectoryIfNotExists_whenExisting_shouldDoNothing() {
        // GIVEN-WHEN
        FilesHelper.createDirectoryIfNotExists(tempDirectory);

        // THEN
        File actualDirectory = new File(tempDirectory);
        assertThat(actualDirectory).exists();
        assertThat(actualDirectory).isDirectory();
    }

    @Test
    public void createDirectoryIfNotExists_whenNonExisting_shouldCreateIt() {
        // GIVEN
        String directoryToCreate = tempDirectory + "/1/2/3";

        // WHEN
        FilesHelper.createDirectoryIfNotExists(directoryToCreate);

        // THEN
        File actualDirectory = new File(directoryToCreate);
        assertThat(actualDirectory).exists();
        assertThat(actualDirectory).isDirectory();
    }
}